src/server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
attrs==22.1.0; python_version >= '3.5'
certifi==2022.9.24; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
distlib==0.3.6
django-cors-headers===3.13.0
django-environ-2==2.3.0
django-environ==0.9.0; python_version >= '3.4' and python_version < '4.0'
django-ipware==4.0.2
django-request-logging==0.7.5
django-split-settings==1.2.0
django===3.2.15
djangorestframework==3.13.1
docutils==0.19; python_version >= '3.7'
drf-spectacular==0.22.1
faker==15.1.0
filelock==3.8.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.16.0; python_version >= '3.7'
lockfile==0.12.2
pip==22.2.2; python_version >= '3.7'
pipenv===2022.1.8
platformdirs==2.5.2; python_version >= '3.7'
psutil==5.9.2
psycopg2===2.9.3
pyrsistent==0.18.1; python_version >= '3.7'
python-daemon==2.3.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.4
pyyaml==6.0; python_version >= '3.6'
requests==2.28.1
setuptools==65.4.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4.0'
uwsgi===2.0.20
virtualenv-clone==0.5.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.16.5; python_version >= '3.6'
